+++ src/main/java/com/tbdev/teaneckminyanim/service/feedback/FeedbackConfigurationException.java
@@ -1,0 +1,7 @@
+package com.tbdev.teaneckminyanim.service.feedback;
+
+public class FeedbackConfigurationException extends Exception {
+    public FeedbackConfigurationException(String message) {
+        super(message);
+    }
+}