+++ src/main/java/com/tbdev/teaneckminyanim/service/feedback/FeedbackValidationException.java
@@ -1,0 +1,7 @@
+package com.tbdev.teaneckminyanim.service.feedback;
+
+public class FeedbackValidationException extends Exception {
+    public FeedbackValidationException(String message) {
+        super(message);
+    }
+}